project/requirements.txt
djangorestframework
Django
Pillow
xlsxwriter
selenium
webdriver-manager
PyQt5
pandas
requests